Fix ARM64 builds
